Add new framework for setting PS1 and the window title.
This commit breaks the prompt; subsequent commits will re-add
functionality. The goals of the new framework are:

Split each part of the prompt into a different file for better
organization.

Switch from printing parts of the prompt in subshells, to appending to
variables in the main shell. This makes it much easier to put untrusted
text into variables to prevent them from being interpreted in PS1 by the
shell.

Instead of doing multiple passes over some components for the prompt and
the window title, calculate both in a single pass. Hopefully this will
make the prompt a bit more responsive.