Reorganize shell startup files.
This commit creates ~/.config/shell/{interactive,login,xlogin}.d/ for
the startup files for any sort of shell startup. Then it sets up other
startup files to source from the appropriate directories.

The minor changes of adding `-g` to some `declare` calls is because the
files are now sourced from inside a function, making `-g` necessary for
the variables to be visible outside that function.