Wrap at 80 columns by default.
Setting textwidth to 80 globally sets the default when the filetype is
unknown, and for many filetypes that don't override textwidth. However,
a few filetype plugins have a per-filetype textwidth, but defer to the
default. This commit sets textwidth globally, and also locally overrides
it to the per-filetype values for filetypes that defer to the global
value.

Command to find such ftplugins (this might not find them all though):

  grep -r 'if.*\<\(tw\|textwidth\)\>' /usr/share/vim/vim80/ftplugin