Make TERM detection more thorough.
In tmux, just set TERM to the best option whether its available or not.

In interactive.d, try to find the best available terminfo for the
initial $TERM, regardless of where that $TERM came from. This should
cover the tmux use case, as well as making sure $TERM is reasonable when
ssh'ing into a machine with a different terminfo database.