Enable the direxpand option.
Previously, variables were getting escaped on expansion, e.g.:

  $ foo=foo
  $ ls $foo/<TAB>
  $ ls \$foo/

That makes it pretty annoying to use variables to store directory paths.
With this change, the expansion is instead:

  $ foo=foo
  $ ls $foo/<TAB>
  $ ls foo/

I'd prefer to keep the variable unexpanded, but I think expanding the
variable is probably going to be less annoying than escaping the dollar
sign.