Make screenshots interactive instead of saving to disk.
Ideally, I'd like to take the screenshot as normal with these
keybindings, but then show a dialog for what to do with the screenshot
(save, copy to clipboard, or discard). I don't see a way to do that, so
the next best thing is to replace the save-to-disk shortcuts with a
single totally interactive shortcut.