Launch ssh in tmux windows instead of terminal tabs
This makes it easier to mix launching individual ssh commands manually
with launching many at once, since it can query which windows already
exist and not create duplicates. It also makes it possible to have a
shell as the parent of ssh so that when ssh exits, I can (e.g.) run ping
to wait for the host to come back online.

This also changes the name of the script because the default is now to
launch a single ssh rather than multiple sshes, and because it makes it
easier to tab complete `ssh foo<Tab>` then add `-l<Tab>` to the command
to tab complete the rest. (I could write a custom completion script, but
it doesn't seem worth it at the moment.)