Only use colorcolumn in modifiable buffers
I had previously disabled it using textwidth in man buffers, but I also
don't want colorcolumn in nomodifiable help buffers, but do want it in
modifiable help buffers.

I'm a bit unsure if using set(local) inside of an OptionSet autocmd is a
good idea, but it seems to work. See
https://github.com/vim/vim/issues/16841 for details.