Make most matches conditional on 'modifiable' instead of filetype
Similar to in 71a5f74 (Only use colorcolumn in modifiable buffers,
2025-03-14), I think it's actually the modifiable setting that's
relevant to whether I want these matches or not, not the filetype. E.g.,
if I'm editing a vim help file, I want the matches, but if I'm reading
one with the :help command, I don't want most of them.

The rfc2119Keyword match is an exception, because it's probably somewhat
useful while reading nomodifiable documentation.