Switch back to gnome-terminal
This partially reverts 08f4c03 (Switch from gnome-terminal to ptyxis,
2025-01-07)

The security around containers is somewhat concerning:
https://gitlab.gnome.org/chergert/ptyxis/-/issues/384

I originally switched because of a cursor blinking difference, but on
Wayland both terminals blinked when they probably shouldn't.
https://gitlab.gnome.org/GNOME/vte/-/merge_requests/1 should fix the
issue for both of them.

I looked at some other terminals too, but chose not to go with Gnome
Console because of https://gitlab.gnome.org/GNOME/console/-/issues/404
or with blackbox because of
https://gitlab.gnome.org/raggesilver/blackbox/-/issues/392 and
https://gitlab.gnome.org/raggesilver/blackbox/-/issues/398